+++ resources/AnchorHelper.cs
@@ -1,4 +1,4 @@
-// BUNDLED-COPY: 0.11.18
+// BUNDLED-COPY: 0.11.19
 using System;
 using System.Runtime.CompilerServices;
 using Avalonia;